Input refactor export const, add components to index
