+++ src/pagesComponents/forgotPassword/ui/sentEmailModal/SentEmailModal.tsx
@@ -2,8 +2,8 @@
 import { memo } from 'react';
 
 import { Button } from '@/shared/button/Button';
-import { Modal } from '@/shared/modal/Modal';
 import { Typography } from '@/shared/ui';
+import { Modal } from '@/shared/ui';
 
 import s from './SentEmailModal.module.scss';
