+++ src/shared/ui/modal/Modal.module.scss
@@ -1,4 +1,4 @@
-@import 'src/app/styles/mixins';
+@import '../../../app/styles/mixins';
 
 /* reset */
 // button,
