+++ src/shared/ui/modal/Stories.module.scss
